Puts cartesian chart rendering funcs into new file
